alert dialog design finished

+++ app/src/main/java/com/example/quakehistory/OnDialogListener.java
@@ -1,5 +1,5 @@
 package com.example.quakehistory;
 
 public interface OnDialogListener {
-    void onDialogPositiveClick(String mag, float magValue, String ctry);
+    void onDialogPositiveClick(String mag, double magValue, String ctry);
 }
